Remove temp dir deletion statement from make clean
Invoking make charts initializes the clean recipe, resulting in the
deletion of the temp directory, $(BUILD_DIR), which is created when
make is invoked. Since this is the location the helm client is
installed by tools/helm_install.sh, the script fails. This commit
removes the statement that deletes the temp directory created when make
is invoked.

Change-Id: I87d41df979a5a4daeb539bcc1330f126a5a407c7