Fix 404 on log retrieval for retries of a step
Retries of a step may execute on different workers.
Each worker only exposes the logs that were generated by that worker.
When assembling the URL for the step, we need to ensure
that we include the retry number as part of the key lookup for a step,
so that it can use the correct worker reference.

Updated ActionsHelper.get_step to use try_number.
The test is included.

Change-Id: I2f8cf9ed70ce344f53ecdcc4edbec924ba6a00c9