plugins: Add get_releases to Armada base operator
Currently, the `ArmadaGetReleasesOperator` does not save any releases
retrieved from the environment. In order to add future functionality
that invokes Helm tests for all releases, deployed releases must be
cataloged. This commit adds a method, `get_releases`, to the Armada
base operator and updates existing plugins to utilize it. Additionally,
this change moves the retrieval of tiller information before the Armada
client is fetched in order to limit subsequent calls to the
`get_tiller_info` method.

Change-Id: Ib4aaec762d509994ce90460d0854e526280c4592