[FIX] Tune Airflow scheduler to be less aggressive
Tunes the interval between checks for tasks to run to every 5 seconds
instead of a continuous loop. In practice this looks like it should be
somewhere between 6 - 10 times less active against the database, and
likewise spawn the same factor less processes. Additionally tunes and
adds notes about some of the other tuning parameters.

Change-Id: I5e28b4c081c53a553072470ae53302915a90dd1a