tools: add test_release script
For now, this script requires specifying a Helm release to test.

Shipyard does not actually require a Helm release. If a Helm release is
not passed as param to test_site, Shipyard will run all Helm tests. This
script does not support this scenario. This script only supports testing
a single Helm release.

Change-Id: I67eb74d7d97d6460872eb1495194c41f08bae055