Locked Werkzeug package release to 0.16.1
The latest Werkzeug package release 1.0.0, released on February 8,
is not backwards compatible with the earler releases of this package,
which is used in Flask, used by apache airflow.

This causes shipyard make image job fail, with missing import errors
from the Werkzeug library.

This change locks the Werkzeuz package release to the last compatible
release with the apache airflow in shipyard.

Change-Id: I54dad4ccc1858f4d5986c6e8e9fbf8f5d9847158